norm

+++ sources/map_parsing.c
@@ -6,7 +6,7 @@
 /*   By: rimagalh <rimagalh@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/06 17:28:29 by gtretiak          #+#    #+#             */
-/*   Updated: 2025/12/16 12:04:00 by rimagalh         ###   ########.fr       */
+/*   Updated: 2025/12/16 12:09:58 by rimagalh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ int	build_map(char *line, t_scene *scene)
 		if (char_processing(line[i], scene, i))
 			return (1);
 	}
-	if (ft_strlen(line) > scene->map_size.x)
+	if ((int)ft_strlen(line) > scene->map_size.x)
 		scene->map_size.x = ft_strlen(line);
 	scene->map[scene->map_row] = ft_strdup(line);
 	if (!scene->map[scene->map_row])
